requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
curlify==3.0.0
dnspython==2.8.0
email-validator==2.3.0
Faker==37.12.0
git-filter-repo==2.47.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.0
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
